Move the Seafowl test process fixture to the main.rs file
